requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argcomplete==1.12.2
astunparse==1.6.3
audiomentations==0.16.0
audioread==2.1.9
av==8.0.3
cachetools==4.2.1
carla==0.9.8
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clean-fid==0.1.13
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
dominate==2.6.0
dpam==0.0.3
et-xmlfile==1.1.0
filelock==3.0.12
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
gdown==3.12.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
joblib==1.0.1
jsonpatch==1.30
jsonpointer==2.0
kaldiio==2.17.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.35.0
lws==1.2.7
Markdown==3.3.4
matplotlib==3.3.4
moviepy==1.0.3
networkx==2.5
nnAudio==0.2.2
numba==0.52.0
numpy==1.21.3
oauthlib==3.1.0
onnxruntime==1.8.0
opencv-python==4.5.1.48
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
Pillow==8.1.1
pooch==1.3.0
proglog==0.1.9
protobuf==3.15.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pyparsing==2.4.7
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
pytorch-fid==0.2.0
pytz==2021.1
PyWavelets==1.1.1
pyworld==0.2.12
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.11.2
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.3.17
toml==0.10.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchfile==0.1.0
torchlibrosa==0.0.9
torchvideotransforms==0.1.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
visdom==0.1.8.9
wav2clip==0.1.0
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
youtube-dl==2021.3.3
yq==2.12.0
